# docs/v1/swaggerv1.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Hanz Car Rentals API",
		"description": "This API is made for Hanz Car Rentals as a school project.",
		"version": "1.2.0",
		"contact": {
			"name": "Hanz",
			"email": "hanz.car.rentals@wolfsoft.solutions",
			"url": "https://hanzcar.rentals/support"
		}
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"responses": {
			"UnauthorizedError": {
				"description": "Unauthorized",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
			"BadRequestError": {
				"description": "Bad request",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
			"NotFoundError": {
				"description": "Not found",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
			"ServerError": {
				"description": "Server error",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			}
		},
		"schemas": {
			"Error": {
				"type": "object",
				"properties": {
					"status": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					}
				}
			}
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	],
	"servers": [
		{
			"url": "https://api.hanzcar.rentals/api/v1/",
			"description": "Production server"
		},
		{
			"url": "http://localhost:55600/api/v1/",
			"description": "Development server"
		}
	],
	"paths": {
...
		"/cars": {
			"get": {
				"summary": "Get all cars",
				"tags": [
					"Cars"
				],
				"description": "This endpoint returns all cars",
				"security": [],
				"responses": {
					"200": {
						"description": "A list of cars",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"brand": {
												"type": "string"
											},
											"model": {
												"type": "string"
											},
											"year": {
												"type": "integer"
											},
											"color": {
												"type": "string"
											},
											"price": {
												"type": "integer"
											},
											"rented_by": {
												"type": "string"
											},
											"licence_plate": {
												"type": "string"
											},
											"seats": {
												"type": "integer"
											},
											"space": {
												"type": "string"
											},
											"transmission": {
												"type": "string"
											},
											"fuel": {
												"type": "string"
											},
											"doors": {
												"type": "integer"
											},
											"towing_weight": {
												"type": "string"
											},
											"maximum_gross_weight": {
												"type": "string"
											},
											"location": {
												"type": "string"
											},
											"image": {
												"type": "string"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/cars/car/{id}": {
			"get": {
				"summary": "Get a car by ID",
				"tags": [
					"Cars"
				],
				"description": "This endpoint returns a car by its ID",
				"security": [],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "A car object",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "integer"
										},
										"brand": {
											"type": "string"
										},
										"model": {
											"type": "string"
										},
										"year": {
											"type": "integer"
										},
										"color": {
											"type": "string"
										},
										"price": {
											"type": "integer"
										},
										"rented_by": {
											"type": "string"
										},
										"licence_plate": {
											"type": "string"
										},
										"seats": {
											"type": "integer"
										},
										"space": {
											"type": "string"
										},
										"transmission": {
											"type": "string"
										},
										"fuel": {
											"type": "string"
										},
										"doors": {
											"type": "integer"
										},
										"towing_weight": {
											"type": "string"
										},
										"maximum_gross_weight": {
											"type": "string"
										},
										"location": {
											"type": "string"
										},
										"image": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/cars/add": {
			"post": {
				"summary": "Add a car",
				"tags": [
					"Cars"
				],
				"description": "This endpoint adds a car",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"brand": {
										"type": "string"
									},
									"model": {
										"type": "string"
									},
									"year": {
										"type": "integer"
									},
									"color": {
										"type": "string"
									},
									"price": {
										"type": "integer"
									},
									"licence_plate": {
										"type": "string"
									},
									"seats": {
										"type": "integer"
									},
									"space": {
										"type": "integer"
									},
									"transmission": {
										"type": "integer"
									},
									"fuel": {
										"type": "string"
									},
									"doors": {
										"type": "integer"
									},
									"towing_weight": {
										"type": "integer"
									},
									"maximum_gross_weight": {
										"type": "integer"
									},
									"location": {
										"type": "string"
									},
									"picture_url": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Car added"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/cars/remove": {
			"delete": {
				"summary": "Remove a car",
				"tags": [
					"Cars"
				],
				"description": "This endpoint removes a car",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"type": "integer"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Car removed"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/cars/update/{id}": {
			"put": {
				"summary": "Update a car",
				"tags": [
					"Cars"
				],
				"description": "This endpoint updates a car",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"color": {
										"type": "string"
									},
									"price": {
										"type": "integer"
									},
									"licence_plate": {
										"type": "string"
									},
									"location": {
										"type": "string"
									},
									"picture_url": {
										"type": "string"
									}
								}
							}
						}
					},
					"responses": {
						"200": {
							"description": "Car updated"
						},
						"401": {
							"$ref": "#/components/responses/UnauthorizedError"
						},
						"403": {
							"$ref": "#/components/responses/UnauthorizedError"
						},
						"404": {
							"$ref": "#/components/responses/UnauthorizedError"
						},
						"500": {
							"$ref": "#/components/responses/UnauthorizedError"
						}
					}
				}
			}
		},
		"/cars/rent/{id}": {
			"put": {
				"summary": "Rent a car",
				"tags": [
					"Cars"
				],
				"description": "This endpoint rents a car",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"user_id": {
										"type": "string"
									},
									"from_date": {
										"type": "string"
									},
									"to_date": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Car rented"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
...
		"/reviews": {
			"get": {
				"summary": "Get all reviews",
				"tags": [
					"Reviews"
				],
				"description": "This endpoint returns all reviews",
				"security": [],
				"responses": {
					"200": {
						"description": "A list of reviews",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"car_id": {
												"type": "integer"
											},
											"user_id": {
												"type": "integer"
											},
											"rating": {
												"type": "integer"
											},
											"comment": {
												"type": "string"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/reviews/add": {
			"post": {
				"summary": "Add a review",
				"tags": [
					"Reviews"
				],
				"description": "This endpoint adds a review",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"car_id": {
										"type": "integer"
									},
									"user_id": {
										"type": "integer"
									},
									"review": {
										"type": "string"
									},
									"rating": {
										"type": "integer"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Review added"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/reviews/remove/{id}": {
			"delete": {
				"summary": "Remove a review",
				"tags": [
					"Reviews"
				],
				"description": "This endpoint removes a review",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Review removed"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		}
	}
}